flask_api/requirements.txt
Flask
google-cloud-bigquery